Renamed files into camel case
