+++ src/app.js
@@ -27,7 +27,7 @@ app.use(limiter)
 
 // API ENtry Point
 // app.use('/api/v1', router)
-const blogRoutes = require('./routes/blog.route.js');
+const blogRoutes = require('./routes/blogRoute.js');
 app.use('/api/v1/blogs', blogRoutes);
 
 // 404 Handler
